trust notebook prep
