+++ src/vs/vscode.proposed.d.ts
@@ -1415,6 +1415,12 @@ declare module 'vscode' {
 		 * The document's current run state
 		 */
 		runState?: NotebookRunState;
+
+		/**
+		 * Whether the document is trusted, default to true
+		 * When false, insecure outputs like HTML, JavaScript, SVG will not be rendered.
+		 */
+		trusted?: boolean;
 	}
 
 	export interface NotebookDocumentContentOptions {
